+++ multiqc/utils/config.py
@@ -80,7 +80,7 @@ with open(searchp_fn) as f:
 # eg. FastQC is usually the first step, so should be last in this list.
 module_order = [
     # Post-alignment analysis results
-    'snpeff', 'qualimap', 'featureCounts', 'methylQA', 'rseqc',
+    'quast', 'snpeff', 'qualimap', 'featureCounts', 'methylQA', 'rseqc',
     'picard', 'preseq', 'samblaster', 'samtools', 'bamtools',
     # Alignment tool stats
     'bismark', 'hicup', 'salmon', 'kallisto', 'star', 'tophat', 'bowtie2', 'bowtie1',