+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### v0.8dev
 Module updates:
+* [**QUAST**](http://quast.bioinf.spbau.ru/) - new module!
+  * QUAST is a tool for assessing de novo assemblies against reference genomes.
 * **Bismark** now supports reports from `bam2nuc`, giving Cytosine coverage in General Stats.
 * **Samtools**
   * `stats` submodule now has separate bar graph showing alignment scores